% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecs.R
\name{as_hori}
\alias{as_hori}
\title{Prints a vector horizontally}
\usage{
as_hori(vec)
}
\arguments{
\item{vec}{a vector}
}
\description{
Prints a vector horizontally
}
\examples{
as_hori(letters)
}
\keyword{as_hori}
