% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecs.R
\name{as_named_vec}
\alias{as_named_vec}
\title{Prints a vector how you would write a vector (except named)}
\usage{
as_named_vec(vec)
}
\arguments{
\item{vec}{a vector}
}
\description{
Prints a vector how you would write a vector (except named)
}
\examples{
as_named_vec(letters)
}
\keyword{as_named_vec}
