% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecs.R
\name{as_vec}
\alias{as_vec}
\title{Prints a vector how you would write a vector}
\usage{
as_vec(vec, name = TRUE)
}
\arguments{
\item{vec}{a vector}

\item{name}{a logical}
}
\description{
Prints a vector how you would write a vector
}
\examples{
as_vec(letters)
}
\keyword{as_vec}
