% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecs.R
\name{as_vert}
\alias{as_vert}
\title{Prints a vector vertically}
\usage{
as_vert(vec)
}
\arguments{
\item{vec}{a vector}
}
\description{
Prints a vector vertically
}
\examples{
as_vert(letters)
}
\keyword{as_vert}
