% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billion.R
\name{billion}
\alias{billion}
\title{Creates a big number}
\usage{
billion(x)
}
\arguments{
\item{x}{a number}
}
\description{
This function creates a billion size number
}
\examples{
billion(1)
}
\keyword{billion}
