% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\docType{data}
\name{gcolors}
\alias{gcolors}
\title{Some custom color palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
gcolors
}
\description{
Tired of ugly default colors in ggplot
}
\examples{
gcolors$p1
}
\keyword{color,}
\keyword{fill}
\keyword{manual}
