% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_merge_stack2.R
\name{gt_merge_stack2}
\alias{gt_merge_stack2}
\title{Stacks two columns in a gt object}
\usage{
gt_merge_stack2(
  gt_object,
  col1,
  col2,
  palette = c("black", "grey"),
  ...,
  small_cap = TRUE,
  font_size = c("14px", "10px"),
  font_weight = c("bold", "bold")
)
}
\arguments{
\item{col1}{the variable on top}

\item{col2}{the variable on bottom}

\item{x}{a gt object}
}
\description{
I found this function on stackoverflow, it's super useful
}
\keyword{gt,}
\keyword{stacking}
