% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{gtheme}
\alias{gtheme}
\title{A custom theme inspired by theme_ipsum}
\usage{
gtheme(base_family = "Arial Narrow", base_size = 14, legend = FALSE)
}
\description{
This function print all rows of a data frame
}
\keyword{ggplot2,}
\keyword{theme}
