% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.R
\name{incp}
\alias{incp}
\title{An increment function}
\usage{
incp(x)
}
\arguments{
\item{x}{integer}
}
\description{
This function increments a counter but prints as well
}
\examples{
x <- 1
incp(x)
}
\keyword{increment}
