% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{Some custom dplyr commands that I use a lot}
\usage{
left(df, var)
}
\description{
front() either var or the last column to the first column
}
\examples{
tibble_display()
}
\keyword{dplyr,}
