% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ls}
\alias{ls}
\title{Makes ls an alias for list.files()}
\format{
An object of class \code{character} of length 9.
}
\usage{
ls
}
\description{
Makes ls an alias for list.files()
}
\keyword{ls}
