% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/million.R
\name{million}
\alias{million}
\title{Creates a big number}
\usage{
million(x = 1)
}
\arguments{
\item{x}{a number}
}
\description{
This function creates a million size number
}
\examples{
million()
}
\keyword{million}
