% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Computes percentages by group}
\usage{
pct(df, ..., .drop = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Some variables.}
}
\value{
data frame
}
\description{
Computes percentages by group
}
\examples{
pct(mtcars, cyl)
}
