% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{points}
\alias{points}
\title{R6 method that keeps tracks of a counter increment and of total points}
\description{
Useful for making quizzes and exams
}
\keyword{points}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-points-count}{\code{points$count()}}
\item \href{#method-points-p}{\code{points$p()}}
\item \href{#method-points-clone}{\code{points$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-points-count"></a>}}
\if{latex}{\out{\hypertarget{method-points-count}{}}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{points$count(x = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-points-p"></a>}}
\if{latex}{\out{\hypertarget{method-points-p}{}}}
\subsection{Method \code{p()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{points$p(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-points-clone"></a>}}
\if{latex}{\out{\hypertarget{method-points-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{points$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
