% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all.R
\name{print_all}
\alias{print_all}
\title{A modified print function}
\usage{
print_all(df)
}
\description{
This function print all rows of a data frame
}
\examples{
print_all(mtcars)
}
\keyword{print}
\keyword{rows}
