% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.R
\name{qp}
\alias{qp}
\title{Keeps track of question points}
\description{
I use this for homeworks because I can assign points
and it will print a formatted number of points as well
as keep track of a running total.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuestionPoints-new}{\code{qp$new()}}
\item \href{#method-QuestionPoints-question}{\code{qp$question()}}
\item \href{#method-QuestionPoints-print}{\code{qp$print()}}
\item \href{#method-QuestionPoints-clone}{\code{qp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qp$new(points = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-question"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-question}{}}}
\subsection{Method \code{question()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qp$question(p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qp$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{qp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
