% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{QuestionPoints}
\alias{QuestionPoints}
\title{A R6 object to keep a running total of homework points.}
\description{
Use for class homeworks.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuestionPoints-new}{\code{QuestionPoints$new()}}
\item \href{#method-QuestionPoints-question}{\code{QuestionPoints$question()}}
\item \href{#method-QuestionPoints-print}{\code{QuestionPoints$print()}}
\item \href{#method-QuestionPoints-clone}{\code{QuestionPoints$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuestionPoints$new(points = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-question"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-question}{}}}
\subsection{Method \code{question()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuestionPoints$question(p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuestionPoints$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuestionPoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuestionPoints-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuestionPoints$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
