% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{mean function with na.rm = TRUE}
\usage{
read_csv(file, ..., show_col_types = FALSE)
}
\description{
Shuts up read_csv's incessant printing of crap
}
\examples{
read_csv(file)
}
\keyword{na.rm}
\keyword{read_csv}
