% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_factors.R
\name{remove_factors}
\alias{remove_factors}
\title{Removes all factors from a data frame}
\usage{
remove_factors(df)
}
\description{
Removes all factors from a data frame
}
\examples{
remove_factors(mtcars)
}
\keyword{factor}
\keyword{killah}
