% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundr.R
\name{roundr}
\alias{roundr}
\title{A rounding function}
\usage{
roundr(x)
}
\arguments{
\item{x}{a number}
}
\description{
custom rounding function: number of digits depends on size of number
}
\examples{
roundr(3.2)
}
\keyword{roundr}
