% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_display.R
\name{tibble_display}
\alias{tibble_display}
\title{A Tibble preference function}
\usage{
tibble_display()
}
\description{
This function makes tibbles wider to fit system environment
}
\examples{
tibble_display()
}
\keyword{options}
\keyword{tibble}
