% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthors.R
\name{time_coauthors}
\alias{time_coauthors}
\title{A timezone function}
\usage{
time_coauthors(time)
}
\description{
This function returns the current time of some of my co-authors
}
\examples{
time_coauthors()
}
\keyword{coauthors}
\keyword{timezone}
