% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pdf.R
\name{to_pdf}
\alias{to_pdf}
\title{Print xaringan slides to PDF
This super useful function was created by Garrick Aden-Buie
www.garrickadenbuie.com/blog/print-xaringan-chromote/}
\usage{
to_pdf(input, output_file = NULL, delay = 1, include_partial_slides = FALSE)
}
\arguments{
\item{input}{Path to Rmd or html file of xaringan slides.}

\item{output_file}{The name of the output file. If using NULL then
the output filename will be based on filename for the input file.
If a filename is provided, a path to the output file can also be provided.}

\item{delay}{Seconds of delay between advancing to and printing
a new slide.}

\item{include_partial_slides}{Should partial (continuation) slides be
included in the output? If \code{FALSE}, the default, only the complete slide
is included in the PDF.}
}
\description{
I only modified the name because I never remember how to spell xaringan
}
\details{
Prints xaringan slides to a PDF file, even complicated slides
with panelsets or other html widgets or advanced features.
Requires a local installation of Chrome.
}
