% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trillion.R
\name{trillion}
\alias{trillion}
\title{Creates a big number}
\usage{
trillion(x = 1)
}
\arguments{
\item{x}{a number}
}
\description{
This function creates a trillion size number
}
\examples{
trillion()
}
\keyword{trillion}
