% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var}
\alias{var}
\title{var function with na.rm = TRUE}
\usage{
var(x, ..., na.rm = TRUE)
}
\arguments{
\item{x}{a number}
}
\description{
Modifies nonsensical default behavior of the var function with respect to NA
}
\examples{
var(c(1,2,3,NA))
}
\keyword{na.rm}
\keyword{var}
