% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_screen.R
\name{wide_screen}
\alias{wide_screen}
\title{A screen function}
\usage{
wide_screen()
}
\description{
This function makes the screen wider given system environment
}
\examples{
howWide=Sys.getenv("COLUMNS")
wide_screen()
}
\keyword{screen}
\keyword{width}
