% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{An additional dplyr verb}
\usage{
left(df, ...)
}
\arguments{
\item{df}{a data frame}
}
\value{
data.frame
}
\description{
This moves either var or the last column to the first column
}
\examples{
left(mtcars)
}
\keyword{dplyr}
