% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\title{mean function with na.rm = TRUE}
\usage{
mean(x, ..., na.rm = TRUE)
}
\description{
Modifies nonsensical default behavior of these functions
}
\examples{
mean(c(1,2,3,NA))
}
\keyword{mean}
\keyword{na.rm}
