% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum}
\alias{sum}
\title{sum function with na.rm = TRUE}
\usage{
sum(x, ..., na.rm = TRUE)
}
\description{
Modifies nonsensical default behavior of these functions
}
\examples{
sum(c(1,2,3,NA))
}
\keyword{na.rm}
\keyword{sum}
